Adds fixes, moves files, and adds aria-label
