+++ app/__tests__/Home.test.tsx
@@ -3,7 +3,7 @@ import { MemoryRouter } from 'react-router-dom';
 import { describe, expect, it } from 'vitest';
 import Hub from '../hub/hub';
 
-describe('Welcome', () => {
+describe('Home', () => {
   it('renders the homepage text', () => {
     render(
       <MemoryRouter>
