+++ app/utils/normalizeMovieResults.ts
@@ -1,8 +1,8 @@
-import type { ListType } from '~/components/list';
-import type { MovieSummary } from './api';
+import type { ListItem } from '~/components/list';
+import type { MovieSummary } from '../api/tmdb_api';
 
 export const normalizeMovieResults = (results: MovieSummary[]) => {
-  const normalizedResults: ListType[] = results.map((item: MovieSummary) => ({
+  const normalizedResults: ListItem[] = results.map((item: MovieSummary) => ({
     id: item.id,
     label: item.title,
     date: item.release_date ?? 'Unknown',
